+++ packages/sdk/src/auth/api/index.ts
@@ -49,11 +49,14 @@ export type {
   CreateSignInResponse,
   ErrorResponse,
   InvalidPasswordResponse,
+  JumpToServiceRequest,
+  JumpToUnidyRequest,
   LoginOptions,
   PasskeyCredential,
   PasskeyOptionsResponse,
   RequiredFieldsResponse,
-  SendMagicCodeError,
+  ResendDelayError,
+  ResendDelayResponse,
   SendMagicCodeResponse,
   TokenResponse,
 } from "./schemas";
@@ -74,6 +77,8 @@ export type {
   GetPasskeyOptionsResult,
   RefreshTokenArgs,
   RefreshTokenResult,
+  ResendConfirmationArgs,
+  ResendConfirmationResult,
   ResetPasswordArgs,
   ResetPasswordResult,
   SendMagicCodeArgs,
@@ -148,6 +153,10 @@ export class AuthService extends BaseService {
 
   signedIn(): Promise<signIn.SignedInResult> {
     return signIn.signedIn(this.client, this.respond);
+  }
+
+  resendConfirmation(args: signIn.ResendConfirmationArgs): Promise<signIn.ResendConfirmationResult> {
+    return signIn.resendConfirmation(this.client, args, this.respond);
   }
 
   // ============================================